refactor: 구입 금액 변수 타입 long을 변경 및 클래스명 변경

+++ src/main/java/lotto/exception/Validator.java
@@ -34,7 +34,7 @@ public class Validator {
         }
     }
 
-    public static void isValidAmount(int amount) {
+    public static void isValidAmount(long amount) {
         if (amount % 1000 != 0) {
             throw new IllegalArgumentException(PURCHASE_AMOUNT_ERROR_MESSAGE);
         }
